requirements.txt
Flask==2.1.2
python-dotenv==0.21.0
google-auth==2.15.0
google-auth-oauthlib==0.7.0
requests==2.28.1
bs4==0.0.1
